+++ packages/build-lib/src/shared.ts
@@ -1,4 +1,4 @@
-import { getAnybuildConfig, getInfo } from 'anybuild'
+import { getsomebuildConfig, getInfo } from 'somebuild'
 
 interface IBuildInfoTsup {
     engine?: 'tsup'
@@ -20,20 +20,20 @@ export function getBuildinfo() {
 }
 
 // overide是否覆盖整个内置的tsup配置
-export type IAnyBuildConfig = {
+export type IsomebuildConfig = {
     overide?: boolean
     bin?: boolean
 } & import('tsup').Options
 /**
  * 定义配置文件
  */
-export function defineConfig(config: IAnyBuildConfig) {
+export function defineConfig(config: IsomebuildConfig) {
     return config
 }
 /**
- * 获取anybuild.config.mts配置文件
+ * 获取somebuild.config.mts配置文件
  */
 export function getConfig() {
-    let config = getAnybuildConfig<IAnyBuildConfig>()
+    let config = getsomebuildConfig<IsomebuildConfig>()
     return config
 }